# docs/swagger.yaml
basePath: /api/v1
definitions:
  controller.emailPassword:
    properties:
      email:
        example: jjh123@naver.com
        type: string
      password:
        example: "123"
        type: string
    type: object
  httputil.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  model.User:
    properties:
      access_key:
        example: cY158XlCRODQljHva8pMjORsoxrKRdfg4S9jT8qa
        type: string
      email:
        example: jjh123@naver.com
        type: string
      name:
        example: jjh
        type: string
      password:
        example: "123"
        type: string
      phone_number:
        example: 010-1234-5678
        type: string
      secret_key:
        example: 2y0BcdVYH48Hxc8SEwfOucxAqMoL623K70j6OCWa
        type: string
    required:
    - access_key
    - email
    - name
    - password
    - phone_number
    - secret_key
    type: object
  model.Wallet:
    properties:
      avg_buy_price:
        type: string
      avg_buy_price_modified:
        type: boolean
      balance:
        type: string
      currency:
        type: string
      locked:
        type: string
      unit_currency:
        type: string
    type: object
  rds.Resp_Quote:
    properties:
      data:
        items:
          properties:
            name:
              type: string
            quote:
              properties:
                BTC:
                  properties:
                    price:
                      type: number
                  type: object
                USD:
                  properties:
                    price:
                      type: number
                  type: object
              type: object
            symbol:
              type: string
          required:
          - name
          - symbol
          type: object
        type: array
      status:
        properties:
          timestamp:
            type: string
        type: object
    type: object
host: localhost:5000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /coin/list:
    get:
      consumes:
      - application/json
      description: get coinlist
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rds.Resp_Quote'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Coin-List
      tags:
      - coin
  /coin/wallet:
    get:
      consumes:
      - application/json
      description: get coinwallet
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Wallet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Coin-Wallet
      tags:
      - coin
  /user/signin:
    post:
      consumes:
      - application/json
      description: 로그인
      operationId: post-user-signin
      parameters:
      - description: User
        in: body
        name: emailPassword
        required: true
        schema:
          $ref: '#/definitions/controller.emailPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Login
      tags:
      - user
  /user/signup:
    post:
      consumes:
      - application/json
      description: 회원가입
      operationId: post-user-signup
      parameters:
      - description: User
        in: body
        name: model.User
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Register
      tags:
      - user
swagger: "2.0"